Don't break if a feed does not have an "rel=alternate type=html" link.
It is possible for a valid Atom feed to omit a

  <link rel="alternate" type="html" href="http://server.com"/>

element which is available through the feedparser object as a

  feed.link

attribute. If the feed does not have this element then the RSS portlet
will throw an AttributeError which will propagate to the page preventing
the original page from rendering.

This changeset adds support for such feeds.